debt - menubar service should not live in common (#79181)
