+++ src/vs/platform/menubar/node/menubar.ts
